pyproject.toml
[project]
name = "ec-hub"
version = "0.1.0"
description = "eBay輸出転売 自動化システム"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "apscheduler>=3.10",
    "aiosqlite>=0.20",
]

[project.scripts]
ec-hub = "ec_hub.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ec_hub"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
